requirements.txt
pyautogui
pytest
pyscreeze
pillow
pytesseract